refactor: merged the errorBoundary component tests into one file
