--- src/_test_/fallback-ui.test.tsx
@@ -1,20 +1,0 @@
-import { render, screen } from '@testing-library/react';
-import { describe, expect, it } from 'vitest';
-import { ErrorBoundary } from '../components/error-boundary/error-boundary';
-
-describe('Displays fallback UI when error occurs', () => {
-  it('Displays fallback UI when error occurs', () => {
-    const ThrowingComponent = () => {
-      throw new Error('Test error');
-    };
-    render(
-      <ErrorBoundary>
-        <ThrowingComponent />
-      </ErrorBoundary>
-    );
-    expect(screen.getByText(/something went wrong/i)).toBeInTheDocument();
-    expect(
-      screen.getByRole('button', { name: /try again/i })
-    ).toBeInTheDocument();
-  });
-});
